refactor: add exception handling decorator
